labs/03_rag_basico/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
pydantic>=2.0.0
streamlit
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=1.3.0 